# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Saylora API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  businessId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  image?: string | null;
  available: boolean;
  createdAt: Date;
}